// actions/core/entryTag/types.ts
export enum EntryTagTypes {
  SET_ENTRY_TAGS = "CORE_ENTRY_TAGS_SET_ENTRY_TAGS",
  ENTRY_TAGS_FETCH_REQUESTED = "CORE_ENTRY_TAGS_FETCH_REQUESTED",
  IS_FETCHING = "CORE_ENTRY_TAGS_IS_FETCHING",
  ARCHIVE_REQUESTED = "CORE_ENTRY_TAGS_ARCHIVE_REQUESTED",
  UNARCHIVE_REQUESTED = "CORE_ENTRY_TAGS_UNARCHIVE_REQUESTED",
  UPDATE_REQUESTED = "CORE_ENTRY_TAGS_UPDATE_REQUESTED",
  CREATE_REQUESTED = "CORE_ENTRY_TAGS_CREATE_REQUESTED",
  CREATED = "CORE_ENTRY_TAGS_CREATED",
  CREATION_ERROR = "CORE_ENTRY_TAGS_CREATION_ERROR",
  CREATION_ERROR_REMOVED = "CORE_ENTRY_TAGS_CREATION_ERROR_REMOVED",
  RESET_ENTRY_TAGS_STATE = "RESET_ENTRY_TAGS_STATE",
}